[graphite] Move YUVImage Precompile factory to public API
This version of the API requires that clients know their future
SkPaint will utilize a YUV Image - which isn't that unreasonable.

An alternative would be to just glom all the YUV combinations into
the PrecompileImageShader and, maybe, add a flag.

This CL is almost entirely cut and paste w/ some comment & header
cleanup.

Bug: b/262606549
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/878460
Commit-Queue: Robert Phillips <robertphillips@google.com>
